% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_ratio.R
\name{cir_ratio}
\alias{cir_ratio}
\title{Theoretical CIR ratio}
\usage{
cir_ratio(mz)
}
\arguments{
\item{mz}{monoisotopic masses (numeric vector).}
}
\value{
CIR theoretical ratios (same length).
}
\description{
Theoretical CIR ratio
}
