% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_adducts.R
\name{default_adducts}
\alias{default_adducts}
\title{Default adduct definitions}
\usage{
default_adducts(ion_mode = c("pos", "neg"))
}
\arguments{
\item{ion_mode}{Ion mode. Either \code{"pos"} or \code{"neg"}. If \code{NULL}, adducts for
both ion modes are returned.}
}
\value{
A data.frame containing adduct definitions, including adduct names,
ion mode and mass shifts relative to the neutral molecule.
}
\description{
Returns the default adduct definitions used by PeakGuideR for the selected
ion mode.
}
\examples{
default_adducts("pos")
default_adducts("neg")
default_adducts(NULL)

}
