% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_pkm.R
\docType{data}
\name{example_pkm}
\alias{example_pkm}
\title{Example PeakGuideR peak matrix}
\format{
A list containing at least:
\describe{
\item{mass}{m/z values of detected features.}
\item{intensity}{Pixel-by-feature intensity matrix.}
\item{pos}{Pixel coordinates.}
}
}
\source{
Internal example dataset.
}
\usage{
data(example_pkm)
}
\description{
Small example peak matrix object used to demonstrate the PeakGuideR workflow.
}
\details{
This dataset is intended for examples, tests and vignette execution. It is a
reduced peak matrix and should not be used for biological interpretation.
}
\keyword{datasets}
