% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_utils.R
\name{gaussian_score}
\alias{gaussian_score}
\title{Gaussian plausibility score from a non-negative error}
\usage{
gaussian_score(err, tol)
}
\arguments{
\item{err}{Numeric vector of non-negative errors.}

\item{tol}{Numeric tolerance.controlling the width of the Gaussian penalty.
With this formulation, the score is approximately 0.61 when \code{err = tol}.}
}
\value{
Numeric vector in range 0 to 1.
}
\description{
This script is designed to make the scores compatible.
}
\keyword{internal}
