% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal_helpers.R
\name{is_cardinal_object}
\alias{is_cardinal_object}
\title{Check whether an object is a supported Cardinal object}
\usage{
is_cardinal_object(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical value.
}
\description{
Internal helper used to detect Cardinal MSI objects before conversion to a
PeakGuideR peak matrix.
}
\keyword{internal}
