% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_peakguider_database.R
\name{load_peakguider_databases}
\alias{load_peakguider_databases}
\title{Load external PeakGuideR databases}
\usage{
load_peakguider_databases(compound_db_path, standards_db_path = NULL)
}
\arguments{
\item{compound_db_path}{Path to \code{compound_mass_database_noncommercial.rds}.}

\item{standards_db_path}{Optional path to
\code{standards_adduct_library_noncommercial.rds}.}
}
\value{
A list with \code{compound_db} and \code{standards_db}.
}
\description{
Loads full PeakGuideR annotation databases downloaded separately, for example
from Zenodo.
}
\examples{
\dontrun{
dbs <- load_peakguider_databases(
  compound_db_path = "path/to/compound_mass_database_noncommercial.rds",
  standards_db_path = "path/to/standards_adduct_library_noncommercial.rds"
)

data(example_pkm, package = "PeakGuideR")

res <- run_peakguider_workflow(
  pkm = example_pkm,
  ion_mode = "pos",
  matrix = "HCCA",
  compound_db = dbs$compound_db,
  standards_db = dbs$standards_db
)
}

}
