% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_peakguider_databases.R
\name{load_standards_adduct_library}
\alias{load_standards_adduct_library}
\title{Load the PeakGuideR standard adduct library}
\usage{
load_standards_adduct_library(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. If \code{FALSE}, prints a message.}
}
\value{
A data.frame.
}
\description{
Loads the example standard adduct library included with PeakGuideR.
The full non-commercial standard adduct library is distributed separately
and can be supplied manually to the workflow through the \code{standards_db}
argument.
}
\examples{
standards_db <- load_standards_adduct_library(quiet = TRUE)
head(standards_db)

}
