% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse molecular formula into element counts}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{Character molecular formula (e.g. "C6H12O6")}
}
\value{
Named numeric vector with element counts
}
\description{
Parse molecular formula into element counts
}
\keyword{internal}
