% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_peakguider_workflow.R
\name{print.peakguider_workflow}
\alias{print.peakguider_workflow}
\title{Print PeakGuideR workflow result}
\usage{
\method{print}{peakguider_workflow}(x, ...)
}
\arguments{
\item{x}{A \code{peakguider_workflow} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print PeakGuideR workflow result
}
