% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinal_helpers.R
\name{cardinal_to_peakmatrix}
\alias{cardinal_to_peakmatrix}
\title{Convert a Cardinal MSImagingExperiment object to a peak matrix}
\usage{
cardinal_to_peakmatrix(
  x,
  value = NULL,
  dataset_name = NULL,
  snr = NULL,
  area = NULL
)
}
\arguments{
\item{x}{A Cardinal MSImagingExperiment object.}

\item{value}{Name of the imageData layer to extract. If NULL, "intensity" is used.}

\item{dataset_name}{Optional dataset name.}

\item{snr}{Optional SNR matrix.}

\item{area}{Optional area matrix.}
}
\value{
An object with rMSIprocPeakMatrix-like structure.
}
\description{
Converts a supported Cardinal MSI object into a PeakGuideR peak matrix with
an rMSIprocPeakMatrix-like structure.
}
\examples{
\dontrun{
# `cardinal_msi_data` is a Cardinal MSImagingExperiment object.
pkm <- cardinal_to_peakmatrix(cardinal_msi_data)

res <- run_peakguider_workflow(
  pkm = pkm,
  ion_mode = "pos",
  matrix = "HCCA"
)
}

}
