% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R, R/package.R
\docType{package}
\name{PeakGuideR-package}
\alias{PeakGuideR-package}
\title{Internal isotope reference tables}
\description{
Internal lookup tables used by the CIR and EIPS scoring functions.

@keywords internal
@importFrom stats approx
@importFrom utils data
}
\details{
\itemize{
\item \code{cir_table}: carbon isotope-ratio reference table.
\item \code{eips_table}: formula-level elemental isotope lookup table.
\item \code{eips_n_table}: theoretical element-count isotope lookup table.
}

These objects are internal implementation data and are not part of the
public PeakGuideR API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nadiamolto/PeakGuideR}
  \item \doi{10.5281/zenodo.20705395}
  \item \url{https://nadiamolto.github.io/PeakGuideR/}
}

}
\author{
\strong{Maintainer}: Nadia Molto \email{nadia.molto@gmail.com} (\href{https://orcid.org/0009-0003-0042-4997}{ORCID})

Authors:
\itemize{
  \item Nadia Molto \email{nadia.molto@gmail.com} (\href{https://orcid.org/0009-0003-0042-4997}{ORCID})
}

}
\keyword{internal}
