% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{PeakGuideR-package}
\alias{PeakGuideR}
\alias{PeakGuideR-package}
\title{PeakGuideR}
\description{
@keywords internal
@importFrom stats approx
@importFrom utils data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nadiamolto/PeakGuideR}
  \item \doi{10.5281/zenodo.20705395}
}

}
\author{
\strong{Maintainer}: Nadia Molto \email{nadia.molto@gmail.com} (\href{https://orcid.org/0009-0003-0042-4997}{ORCID})

Authors:
\itemize{
  \item Nadia Molto \email{nadia.molto@gmail.com} (\href{https://orcid.org/0009-0003-0042-4997}{ORCID})
}

}
